Improve organization by separating headers/sources
